refactor(ui-info-state): suffix every web-only file with .web
The package suffixed some web-only files (types.web.ts, PresetVisual.web.tsx)
but not others (the whole background/ folder, internals/cn.ts, src/index.ts),
even though those are equally web-only. The inconsistency was load-bearing:
because tsconfig.native.json excludes src/**/*.web.*, the unsuffixed files were
root files of the native program, and src/index.ts re-exported the web tree on
top of that, so the native typecheck was compiling web code.

Suffixing them lets the existing *.web.* exclude do the whole job, so the two
path excludes added in the previous commit are no longer needed and are
reverted. The native program is now exactly the native sources plus the
platform-neutral sharedTypes.ts, which is what makes the native typecheck a
real guarantee that native code pulls in no web-only module.

package.json main/types/exports follow src/index.ts to src/index.web.ts;
resolution is unchanged (react-native -> index.native.ts, default ->
index.web.ts).
