+++ shared/ui-info-state/package.json
@@ -3,13 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "description": "Shared InfoState layout for web (Desktop) and React Native (Mobile), plus the web-only dialog background tone plumbing it uses",
-  "main": "src/index.ts",
+  "main": "src/index.web.ts",
   "react-native": "src/index.native.ts",
-  "types": "src/index.ts",
+  "types": "src/index.web.ts",
   "exports": {
     ".": {
       "react-native": "./src/index.native.ts",
-      "default": "./src/index.ts"
+      "default": "./src/index.web.ts"
     },
     "./package.json": "./package.json"
   },
