+++ shared/ui-info-state/src/InfoState.web.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { Search } from "@ledgerhq/lumen-ui-react/symbols";
-import { DialogBackgroundContext } from "./background/DialogBackgroundContext";
+import { DialogBackgroundContext } from "./background/DialogBackgroundContext.web";
 import { InfoState } from "./InfoState.web";
 import type { InfoStatePreset } from "./types.web";
