--- shared/ui-info-state/src/background/index.ts
@@ -1,4 +1,0 @@
-export * from "./DialogBackgroundContext";
-export * from "./useDialogBackgroundTone";
-export * from "./useDialogBackgroundToneRequests";
-export * from "./DialogBackgroundGradient";
