+++ shared/ui-info-state/src/index.web.ts
@@ -1,3 +1,3 @@
 export * from "./InfoState.web";
 export * from "./types.web";
-export * from "./background";
+export * from "./background/index.web";
